feat: Add foundational component interfaces, types, and new DSA-related pages and components.
